+++ packages/expo-sqlite/src/SQLite.ts
@@ -54,6 +54,26 @@ export class SQLiteDatabase {
   }
 
   /**
+   * Due to limitations on `Android` this function is provided to allow raw SQL queries to be
+   * executed on the database. This will be less efficient than using the `exec` function, please use
+   * only when necessary.
+   */
+  execRawQuery(queries: Query[], readOnly: boolean, callback: SQLiteCallback): void {
+    if (Platform.OS === 'ios') {
+      return this.exec(queries, readOnly, callback);
+    }
+
+    ExpoSQLite.execRawQuery(this._name, queries.map(_serializeQuery), readOnly).then(
+      (nativeResultSets) => {
+        callback(null, nativeResultSets.map(_deserializeResultSet));
+      },
+      (error) => {
+        callback(error instanceof Error ? error : new Error(error));
+      }
+    );
+  }
+
+  /**
    * Executes the SQL statement and returns a Promise resolving with the result.
    */
   async execAsync(queries: Query[], readOnly: boolean): Promise<(ResultSetError | ResultSet)[]> {
@@ -207,6 +227,7 @@ export function openDatabase(
   }
   const db = _openExpoSQLiteDatabase(name, version, description, size, callback);
   db.exec = db._db.exec.bind(db._db);
+  db.execRawQuery = db._db.execRawQuery.bind(db._db);
   db.execAsync = db._db.execAsync.bind(db._db);
   db.closeAsync = db._db.closeAsync.bind(db._db);
   db.deleteAsync = db._db.deleteAsync.bind(db._db);